removed header only plugin_utils library
